Fixed some recently introduced problems in RTP parsing
git-svn-id: svn://svn.minisip.org/minisip/trunk@1364 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ minisip/rtp/RtpHeader.cxx
@@ -67,20 +67,20 @@ int RtpHeader::getPayloadType(){
 }
 
 
-void RtpHeader::setSeqNo(int seq_no){
+void RtpHeader::setSeqNo(uint16_t seq_no){
 	this->sequence_number=seq_no;
 }
 
-int RtpHeader::getSeqNo(){
+uint16_t RtpHeader::getSeqNo(){
 	return sequence_number;
 }
 
 
-void RtpHeader::setTimestamp(int timestamp){
+void RtpHeader::setTimestamp(uint32_t timestamp){
 	this->timestamp = timestamp;
 }
 
-int RtpHeader::getTimestamp(){
+uint32_t RtpHeader::getTimestamp(){
 	return timestamp;
 }
